@@ -34,8 +34,8 @@
     "@azure/logger": "1.0.3"
   },
   "peerDependencies": {
-    "react": ">=16.8.0 <18.0.0",
-    "@types/react": ">=16.8.0 <18.0.0"
+    "react": ">=16.8.0 <19.0.0",
+    "@types/react": ">=16.8.0 <19.0.0"
   },
   "devDependencies": {
     "@babel/cli": "~7.16.0",
@@ -43,7 +43,7 @@
     "@microsoft/api-documenter": "~7.12.11",
     "@microsoft/api-extractor": "~7.18.0",
     "@types/jest": "~26.0.22",
-    "@types/react": "^16.9.49",
+    "@types/react": "18.2.0",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
     "@zerollup/ts-transform-paths": "~1.7.18",
@@ -60,7 +60,7 @@
     "if-env": "~1.0.4",
     "jest": "26.6.0",
     "prettier": "2.3.1",
-    "react": "~16.14.0",
+    "react": "18.2.0",
     "rimraf": "^2.6.2",
     "rollup": "~2.42.4",
     "ts-jest": "^26.4.4",
