@@ -39,18 +39,18 @@
     "memoize-one": "~5.2.1"
   },
   "peerDependencies": {
-    "@azure/communication-chat": "1.3.2-beta.1 || >=1.2.0",
-    "@types/react": ">=16.8.0 <18.0.0",
-    "react": ">=16.8.0 <18.0.0"
+    "@azure/communication-chat": "1.3.2-beta.2 || >=1.2.0",
+    "@types/react": ">=16.8.0 <19.0.0",
+    "react": ">=16.8.0 <19.0.0"
   },
   "devDependencies": {
-    "@azure/communication-chat": "1.3.2-beta.1 || >=1.2.0",
+    "@azure/communication-chat": "1.3.2-beta.2 || >=1.2.0",
     "@babel/cli": "~7.16.0",
     "@babel/core": "~7.16.0",
     "@microsoft/api-documenter": "~7.12.11",
     "@microsoft/api-extractor": "~7.18.0",
     "@types/jest": "~26.0.22",
-    "@types/react": "^16.9.49",
+    "@types/react": "18.2.0",
     "@zerollup/ts-transform-paths": "~1.7.18",
     "copyfiles": "^2.4.1",
     "cpx": "~1.5.0",
@@ -65,7 +65,7 @@
     "if-env": "~1.0.4",
     "jest": "26.6.0",
     "prettier": "2.3.1",
-    "react": "~16.14.0",
+    "react": "18.2.0",
     "rimraf": "^2.6.2",
     "rollup": "~2.42.4",
     "typescript": "4.3.5"
