@@ -51,11 +49,11 @@
   "peerDependencies": {
     "@azure/communication-calling": "1.13.2-beta.1 || >=1.13.1",
     "@azure/communication-calling-effects": "1.0.0-beta.2",
-    "@azure/communication-chat": "1.3.2-beta.1 || >=1.2.0",
-    "@types/react": ">=16.8.0 <18.0.0",
-    "@types/react-dom": ">=16.8.0 <18.0.0",
-    "react": ">=16.8.0 <18.0.0",
-    "react-dom": ">=16.8.0 <18.0.0"
+    "@azure/communication-chat": "1.3.2-beta.2 || >=1.2.0",
+    "@types/react": ">=16.8.0 <19.0.0",
+    "@types/react-dom": ">=16.8.0 <19.0.0",
+    "react": ">=16.8.0 <19.0.0",
+    "react-dom": ">=16.8.0 <19.0.0"
   },
   "main": "./dist/dist-cjs/communication-react/index.js",
   "module": "./dist/dist-esm/communication-react/src/index.js",
@@ -136,10 +134,9 @@
     "node-forge": ">=1.0.0",
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
-    "react-dom": "16.13.1",
+    "react-dom": "18.2.0",
     "react-is": "~17.0.1",
-    "react-test-renderer": "^16.14.0",
-    "react": "~16.14.0",
+    "react": "18.2.0",
     "regenerator-runtime": "^0.13.7",
     "renamer": "~4.0.0",
     "rimraf": "^2.6.2",
