@@ -3,36 +3,43 @@
   "version": "1.5.1-beta.4",
   "private": "true",
   "description": "Package to wrap the Fluent NorthStar UI library, for short-term React 18 compatibility",
+  "module": "dist/dist-esm/index.js",
+  "main": "dist/dist-cjs/index.js",
+  "types": "dist/northstar-wrapper.d.ts",
   "scripts": {
-    "build": "tsc",
+    "build": "rushx build:esm && rushx build:cjs",
+    "build:esm": "tsc && webpack -c ./webpack.esm.config.js && rushx api-extractor",
+    "build:cjs": "webpack -c ./webpack.cjs.config.js",
     "build:watch": "echo skip",
-    "build:stable": "tsc",
+    "build:stable": "tsc && webpack",
+    "api-extractor": "api-extractor run --local",
     "prettier": "prettier --no-error-on-unmatched-pattern --write --config ../../.prettierrc --ignore-path=../../.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
     "prettier:check": "prettier --no-error-on-unmatched-pattern --check --config ../../.prettierrc --ignore-path=../../.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
     "test": "echo skip",
     "test:stable": "echo skip",
     "test:coverage": "echo skip",
-    "lint": "eslint --max-warnings 0 \"**/*.{js,ts,tsx}\"",
+    "lint": "eslint --max-warnings 0 \"./src/**/*.{js,ts,tsx}\"",
     "lint:fix": "rushx lint -- --fix",
     "lint:quiet": "rushx lint -- --quiet"
   },
-  "dependencies": {
-    "@fluentui/react-northstar": "0.61.0",
-    "@fluentui/react": "~8.98.3"
-  },
   "devDependencies": {
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
+    "@fluentui/react-northstar": "0.61.0",
+    "@microsoft/api-extractor": "~7.18.0",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.1",
+    "@typescript-eslint/parser": "^4.12.0",
     "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-header": "^3.1.0",
     "eslint-plugin-import": "~2.22.1",
     "prettier": "2.3.1",
-    "react": "~16.14.0",
-    "react-dom": "16.13.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "scheduler": "^0.20.0",
     "rimraf": "^2.6.2",
-    "typescript": "3.8.3"
+    "typescript": "3.8.3",
+    "webpack": "5.76.0",
+    "webpack-cli": "~4.10.0"
   }
 }
