@@ -38,14 +38,11 @@
     "@fluentui/react-file-type-icons": "8.8.13",
     "@fluentui/react-hooks": "~8.6.12",
     "@fluentui/react-icons": "~2.0.194",
-    "@fluentui/react-northstar": "0.61.0",
     "@fluentui/react-window-provider": "~2.2.3",
-    "@fluentui/react": "~8.98.3",
     "@internal/acs-ui-common": "1.5.1-beta.4",
     "@internal/northstar-wrapper": "1.5.1-beta.4",
     "copy-to-clipboard": "~3.3.1",
     "html-to-react": "~1.5.0",
-    "react-aria-live": "^2.0.5",
     "react-linkify": "^1.0.0-alpha",
     "react-use-draggable-scroll": "~0.4.7",
     "textarea-caret-ts": "^4.1.1",
@@ -53,10 +50,10 @@
     "uuid": "^8.1.0"
   },
   "peerDependencies": {
-    "@types/react": ">=16.8.0 <18.0.0",
-    "@types/react-dom": ">=16.8.0 <18.0.0",
-    "react": ">=16.8.0 <18.0.0",
-    "react-dom": ">=16.8.0 <18.0.0"
+    "@types/react": ">=16.8.0 <19.0.0",
+    "@types/react-dom": ">=16.8.0 <19.0.0",
+    "react": ">=16.8.0 <19.0.0",
+    "react-dom": ">=16.8.0 <19.0.0"
   },
   "devDependencies": {
     "@babel/cli": "~7.16.0",
@@ -66,15 +63,13 @@
     "@microsoft/api-extractor": "~7.18.0",
     "@rollup/plugin-json": "~4.1.0",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "12.1.5",
-    "@testing-library/react-hooks": "^3.4.2",
+    "@testing-library/react": "14.0.0",
     "@types/jest": "~26.0.22",
     "@types/json-stringify-safe": "^5.0.0",
-    "@types/node": "^14.14.10",
-    "@types/react-aria-live": "^2.0.0",
-    "@types/react-dom": "^16.9.8",
+    "@types/node": "^16.11.7",
+    "@types/react-dom": "18.2.1",
     "@types/react-linkify": "^1.0.0",
-    "@types/react": "^16.9.49",
+    "@types/react": "18.2.0",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
@@ -107,10 +102,9 @@
     "node-forge": ">=1.0.0",
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
-    "react-dom": "16.13.1",
+    "react-dom": "18.2.0",
     "react-is": "~17.0.1",
-    "react-test-renderer": "^16.14.0",
-    "react": "~16.14.0",
+    "react": "18.2.0",
     "regenerator-runtime": "^0.13.7",
     "rimraf": "^2.6.2",
     "rollup": "~2.42.4",
