@@ -66,7 +66,6 @@
     "immer": "9.0.6",
     "memoize-one": "~5.2.1",
     "nanoid": "3.1.32",
-    "react-aria-live": "^2.0.5",
     "react-linkify": "^1.0.0-alpha",
     "reselect": "~4.0.0",
     "uuid": "^8.1.0"
@@ -74,18 +73,18 @@
   "peerDependencies": {
     "@azure/communication-calling": "1.13.2-beta.1 || >=1.13.1",
     "@azure/communication-calling-effects": "1.0.0-beta.2",
-    "@azure/communication-chat": "1.3.2-beta.1 || >=1.2.0",
-    "@types/react": ">=16.8.0 <18.0.0",
-    "@types/react-dom": ">=16.8.0 <18.0.0",
-    "react": ">=16.8.0 <18.0.0",
-    "react-dom": ">=16.8.0 <18.0.0"
+    "@azure/communication-chat": "1.3.2-beta.2 || >=1.2.0",
+    "@types/react": ">=16.8.0 <19.0.0",
+    "@types/react-dom": ">=16.8.0 <19.0.0",
+    "react": ">=16.8.0 <19.0.0",
+    "react-dom": ">=16.8.0 <19.0.0"
   },
   "devDependencies": {
     "@azure/communication-calling": "1.13.2-beta.1 || >=1.13.1",
     "@azure/communication-calling-effects": "1.0.0-beta.2",
-    "@azure/communication-chat": "1.3.2-beta.1 || >=1.2.0",
+    "@azure/communication-chat": "1.3.2-beta.2 || >=1.2.0",
     "@azure/communication-identity": "~1.1.0",
-    "@azure/communication-signaling": "1.0.0-beta.17",
+    "@azure/communication-signaling": "1.0.0-beta.18",
     "@babel/cli": "~7.16.0",
     "@babel/core": "~7.16.0",
     "@babel/preset-env": "7.13.10",
@@ -94,18 +93,16 @@
     "@microsoft/api-extractor": "~7.18.0",
     "@playwright/test": "~1.22.2",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "12.1.5",
-    "@testing-library/react-hooks": "^3.4.2",
+    "@testing-library/react": "14.0.0",
     "@types/copy-webpack-plugin": "^6.4.0",
     "@types/express": "^4.17.8",
     "@types/jest": "~26.0.22",
     "@types/json-stringify-safe": "^5.0.0",
-    "@types/node": "^14.14.10",
+    "@types/node": "^16.11.7",
     "@types/puppeteer": "~5.4.3",
-    "@types/react-aria-live": "^2.0.0",
-    "@types/react-dom": "^16.9.8",
+    "@types/react-dom": "18.2.1",
     "@types/react-linkify": "^1.0.0",
-    "@types/react": "^16.9.49",
+    "@types/react": "18.2.0",
     "@types/uuid": "^8.3.0",
     "@types/yargs": "17.0.10",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
@@ -147,10 +144,9 @@
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
     "puppeteer": "~10.0.0",
-    "react-dom": "16.13.1",
+    "react-dom": "18.2.0",
     "react-is": "~17.0.1",
-    "react-test-renderer": "^16.14.0",
-    "react": "~16.14.0",
+    "react": "18.2.0",
     "regenerator-runtime": "^0.13.7",
     "rimraf": "^2.6.2",
     "rollup": "~2.42.4",
