@@ -19,7 +19,7 @@
   },
   "dependencies": {
     "@azure/communication-calling": "1.13.2-beta.1 || >=1.13.1",
-    "@azure/communication-chat": "1.3.2-beta.1 || >=1.2.0",
+    "@azure/communication-chat": "1.3.2-beta.2 || >=1.2.0",
     "@azure/communication-common": "2.2.0",
     "@azure/communication-identity": "~1.1.0",
     "@azure/communication-react": "1.5.1-beta.4",
@@ -61,15 +61,13 @@
     "@storybook/storybook-deployer": "~2.8.16",
     "@storybook/theming": "~6.5.14",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react-hooks": "^3.4.2",
     "@types/classnames": "^2.2.11",
     "@types/jest": "~26.0.22",
     "@types/json-stringify-safe": "^5.0.0",
-    "@types/node": "^14.14.10",
-    "@types/react-aria-live": "^2.0.0",
-    "@types/react-dom": "^16.9.8",
+    "@types/node": "^16.11.7",
+    "@types/react-dom": "18.2.1",
     "@types/react-linkify": "^1.0.0",
-    "@types/react": "^16.9.49",
+    "@types/react": "18.2.0",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
@@ -99,10 +97,9 @@
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
     "raw-loader": "~4.0.2",
-    "react-dom": "16.13.1",
+    "react-dom": "18.2.0",
     "react-is": "~17.0.1",
-    "react-test-renderer": "^16.14.0",
-    "react": "~16.14.0",
+    "react": "18.2.0",
     "regenerator-runtime": "^0.13.7",
     "require-from-string": "^2.0.2",
     "scheduler": "^0.20.0",
