@@ -2,10 +2,11 @@
 // Licensed under the MIT license.
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { AzureCommunicationTokenCredential } from '@azure/communication-common';
 import { CallWithChatComposite, createAzureCommunicationCallWithChatAdapter } from '@azure/communication-react';
-
+import { initializeIcons } from '@fluentui/react';
+initializeIcons();
 // locator is a different type of custom locator to aggregate a call locator and a chat thread
 // locator = { callLocator: CallLocator, chatThreadId : string }
 export const loadCallWithChatComposite = async function (args, htmlElement, props) {
@@ -17,6 +18,11 @@ export const loadCallWithChatComposite = async function (args, htmlElement, prop
     endpoint: endpoint,
     locator: locator
   });
-  ReactDOM.render(React.createElement(CallWithChatComposite, { ...props, adapter }, null), htmlElement);
+
+  if (!htmlElement) {
+    throw new Error('Failed to find the root element');
+  }
+
+  createRoot(htmlElement).render(React.createElement(CallWithChatComposite, { ...props, adapter }, null));
   return adapter;
 };
