@@ -1,14 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import { Stack, FocusZone } from '@fluentui/react';
+import { Stack, FocusZone, Spinner } from '@fluentui/react';
 import React, { useEffect, useRef, useState, useCallback } from 'react';
 import { _FileUploadCardsStrings } from './FileUploadCards';
-import { Ref } from '@fluentui/react-northstar';
+import { Ref } from '@internal/northstar-wrapper';
 import { _Caption } from './Caption';
 import {
   captionContainerClassName,
   captionsBannerClassName,
-  captionsContainerClassName
+  captionsContainerClassName,
+  loadingBannerClassName
 } from './styles/Captions.style';
 import { OnRenderAvatarCallback } from '../types';
 
@@ -25,17 +26,27 @@ export type _CaptionsInfo = {
 
 /**
  * @internal
+ * strings for captions banneer
+ */
+export interface _CaptionsBannerStrings {
+  captionsBannerSpinnerText?: string;
+}
+
+/**
+ * @internal
  * _CaptionsBanner Component Props.
  */
 export interface _CaptionsBannerProps {
   captions: _CaptionsInfo[];
   isCaptionsOn?: boolean;
+  startCaptionsInProgress?: boolean;
   /**
    * Optional callback to override render of the avatar.
    *
    * @param userId - user Id
    */
   onRenderAvatar?: OnRenderAvatarCallback;
+  strings?: _CaptionsBannerStrings;
 }
 
 /**
@@ -43,7 +54,7 @@ export interface _CaptionsBannerProps {
  * A component for displaying a CaptionsBanner with user icon, displayName and captions text.
  */
 export const _CaptionsBanner = (props: _CaptionsBannerProps): JSX.Element => {
-  const { captions, isCaptionsOn, onRenderAvatar } = props;
+  const { captions, isCaptionsOn, startCaptionsInProgress, onRenderAvatar, strings } = props;
   const captionsScrollDivRef = useRef<HTMLElement>(null);
   const [isAtBottomOfScroll, setIsAtBottomOfScroll] = useState<boolean>(true);
 
@@ -82,19 +93,26 @@ export const _CaptionsBanner = (props: _CaptionsBannerProps): JSX.Element => {
 
   return (
     <>
-      {isCaptionsOn && (
+      {startCaptionsInProgress && (
         <FocusZone as="ul" className={captionsContainerClassName}>
-          <Ref innerRef={captionsScrollDivRef}>
-            <Stack verticalAlign="start" className={captionsBannerClassName}>
-              {captions.map((caption) => {
-                return (
-                  <div key={caption.id} className={captionContainerClassName} data-is-focusable={true}>
-                    <_Caption {...caption} onRenderAvatar={onRenderAvatar} />
-                  </div>
-                );
-              })}
+          {isCaptionsOn && (
+            <Ref innerRef={captionsScrollDivRef}>
+              <Stack verticalAlign="start" className={captionsBannerClassName}>
+                {captions.map((caption) => {
+                  return (
+                    <div key={caption.id} className={captionContainerClassName} data-is-focusable={true}>
+                      <_Caption {...caption} onRenderAvatar={onRenderAvatar} />
+                    </div>
+                  );
+                })}
+              </Stack>
+            </Ref>
+          )}
+          {!isCaptionsOn && (
+            <Stack verticalAlign="center" className={loadingBannerClassName} data-is-focusable={true}>
+              <Spinner label={strings?.captionsBannerSpinnerText} ariaLive="assertive" labelPosition="right" />
             </Stack>
-          </Ref>
+          )}
         </FocusZone>
       )}
     </>
