@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 import { concatStyleSets, Icon, ITextField, mergeStyles, Stack } from '@fluentui/react';
+import { Chat } from '@internal/northstar-wrapper';
 import { _formatString } from '@internal/acs-ui-common';
 import { useTheme } from '../../theming/FluentThemeProvider';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
@@ -13,7 +14,7 @@ import { borderAndBoxShadowStyle } from '../styles/SendBox.styles';
 import { ChatMessage } from '../../types';
 import { _FileUploadCards } from '../FileUploadCards';
 import { FileMetadata } from '../FileDownloadCards';
-import { chatMessageFailedTagStyle } from '../styles/ChatMessageComponent.styles';
+import { chatMessageFailedTagStyle, chatMessageEditContainerStyle } from '../styles/ChatMessageComponent.styles';
 /* @conditional-compile-remove(mention) */
 import { MentionLookupOptions } from '../MentionPopover';
 
@@ -114,70 +115,74 @@ export const ChatMessageComponentAsEditBox = (props: ChatMessageComponentAsEditB
     );
   }, [attachedFilesMetadata]);
 
-  return (
-    <Stack
-      className={mergeStyles(
-        borderAndBoxShadowStyle({
-          theme,
-          hasErrorMessage: message.failureReason !== undefined,
-          disabled: false
-        })
-      )}
-    >
-      <InputBoxComponent
-        inlineChildren={props.inlineEditButtons}
-        id={'editbox'}
-        textFieldRef={editTextFieldRef}
-        inputClassName={editBoxStyle(props.inlineEditButtons)}
-        placeholderText={strings.editBoxPlaceholderText}
-        textValue={textValue}
-        onChange={setText}
-        onEnterKeyDown={() => {
-          submitEnabled &&
-            onSubmit(textValue, message.metadata, {
-              attachedFilesMetadata
-            });
-        }}
-        supportNewline={false}
-        maxLength={MAXIMUM_LENGTH_OF_MESSAGE}
-        errorMessage={textTooLongMessage}
-        styles={editBoxStyles}
-        /* @conditional-compile-remove(mention) */
-        mentionLookupOptions={mentionLookupOptions}
+  const getContent = (): JSX.Element => {
+    return (
+      <Stack
+        className={mergeStyles(
+          borderAndBoxShadowStyle({
+            theme,
+            hasErrorMessage: message.failureReason !== undefined,
+            disabled: false
+          })
+        )}
       >
-        <InputBoxButton
-          className={editingButtonStyle}
-          ariaLabel={strings.editBoxCancelButton}
-          tooltipContent={strings.editBoxCancelButton}
-          onRenderIcon={onRenderThemedCancelIcon}
-          onClick={() => {
-            onCancel && onCancel(message.messageId);
-          }}
-          id={'dismissIconWrapper'}
-        />
-        <InputBoxButton
-          className={editingButtonStyle}
-          ariaLabel={strings.editBoxSubmitButton}
-          tooltipContent={strings.editBoxSubmitButton}
-          onRenderIcon={onRenderThemedSubmitIcon}
-          onClick={(e) => {
+        <InputBoxComponent
+          inlineChildren={props.inlineEditButtons}
+          id={'editbox'}
+          textFieldRef={editTextFieldRef}
+          inputClassName={editBoxStyle(props.inlineEditButtons)}
+          placeholderText={strings.editBoxPlaceholderText}
+          textValue={textValue}
+          onChange={setText}
+          onEnterKeyDown={() => {
             submitEnabled &&
               onSubmit(textValue, message.metadata, {
                 attachedFilesMetadata
               });
-            e.stopPropagation();
           }}
-          id={'submitIconWrapper'}
-        />
-      </InputBoxComponent>
-      {message.failureReason && (
-        <div className={mergeStyles(chatMessageFailedTagStyle(theme), { padding: '0.5rem' })}>
-          {message.failureReason}
-        </div>
-      )}
-      {onRenderFileUploads()}
-    </Stack>
-  );
+          supportNewline={false}
+          maxLength={MAXIMUM_LENGTH_OF_MESSAGE}
+          errorMessage={textTooLongMessage}
+          styles={editBoxStyles}
+          /* @conditional-compile-remove(mention) */
+          mentionLookupOptions={mentionLookupOptions}
+        >
+          <InputBoxButton
+            className={editingButtonStyle}
+            ariaLabel={strings.editBoxCancelButton}
+            tooltipContent={strings.editBoxCancelButton}
+            onRenderIcon={onRenderThemedCancelIcon}
+            onClick={() => {
+              onCancel && onCancel(message.messageId);
+            }}
+            id={'dismissIconWrapper'}
+          />
+          <InputBoxButton
+            className={editingButtonStyle}
+            ariaLabel={strings.editBoxSubmitButton}
+            tooltipContent={strings.editBoxSubmitButton}
+            onRenderIcon={onRenderThemedSubmitIcon}
+            onClick={(e) => {
+              submitEnabled &&
+                onSubmit(textValue, message.metadata, {
+                  attachedFilesMetadata
+                });
+              e.stopPropagation();
+            }}
+            id={'submitIconWrapper'}
+          />
+        </InputBoxComponent>
+        {message.failureReason && (
+          <div className={mergeStyles(chatMessageFailedTagStyle(theme), { padding: '0.5rem' })}>
+            {message.failureReason}
+          </div>
+        )}
+        {onRenderFileUploads()}
+      </Stack>
+    );
+  };
+
+  return <Chat.Message styles={chatMessageEditContainerStyle} content={getContent()} />;
 };
 
 const isMessageTooLong = (messageText: string): boolean => messageText.length > MAXIMUM_LENGTH_OF_MESSAGE;
